requirements.txt
streamlit>=1.33
pillow>=10.0.0
pydicom>=2.4.0
numpy>=1.24.0
pylibjpeg>=1.4.0
pylibjpeg-libjpeg>=1.4.0
pylibjpeg-openjpeg>=1.4.0
pdf2image>=1.17.0
